[REF/#247] core.common.event 패키지를 provider와 trigger로 분리
