--- core/common/src/main/java/com/hilingual/core/common/event/trigger/LocalSnackbarTrigger.kt
@@ -1,7 +1,0 @@
-package com.hilingual.core.common.event.trigger
-
-import androidx.compose.runtime.staticCompositionLocalOf
-
-val LocalSnackbarTrigger = staticCompositionLocalOf<() -> Unit> {
-    error("No SnackBar provided")
-}
