+++ core/common/src/main/java/com/hilingual/core/common/provider/SystemBarsColor.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hilingual.core.common.event.provider
+package com.hilingual.core.common.provider
 
 import android.app.Activity
 import android.os.Build
