+++ core/designsystem/src/main/java/com/hilingual/core/designsystem/component/dialog/HIlingualErrorDialog.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.DialogProperties
-import com.hilingual.core.common.event.trigger.DialogState
+import com.hilingual.core.common.trigger.DialogState
 import com.hilingual.core.designsystem.R
 import com.hilingual.core.designsystem.theme.HilingualTheme
