+++ core/designsystem/src/main/java/com/hilingual/core/designsystem/theme/Theme.kt
@@ -24,7 +24,7 @@ import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.ReadOnlyComposable
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.staticCompositionLocalOf
-import com.hilingual.core.common.event.provider.LocalSharedTransitionScope
+import com.hilingual.core.common.provider.LocalSharedTransitionScope
 
 private val LocalHilingualColors = staticCompositionLocalOf<HilingualColors> {
     error("No HilingualColors provided")
