+++ presentation/main/src/main/java/com/hilingual/presentation/main/MainActivity.kt
@@ -22,8 +22,8 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.runtime.CompositionLocalProvider
-import com.hilingual.core.common.event.provider.LocalSystemBarsColor
-import com.hilingual.core.common.event.provider.SystemBarsColorController
+import com.hilingual.core.common.provider.LocalSystemBarsColor
+import com.hilingual.core.common.provider.SystemBarsColorController
 import com.hilingual.core.designsystem.theme.HilingualTheme
 import com.hilingual.presentation.main.monitor.NetworkMonitor
 import com.hilingual.presentation.main.state.rememberMainAppState
